jira_manager_2/requirements.txt
jira>=3.5.0
requests>=2.28.0
python-dotenv>=1.0.0
